Prep and support for fallback flag
