+++ gulpfile.js
@@ -2,7 +2,6 @@
 'use strict';
 
 const gulp = require('gulp');
-const gutil = require('gulp-util');
 const ignore = require('gulp-ignore');
 const closureBuild = require('closure-builder');
 const argv = require('yargs').argv;
@@ -22,6 +21,7 @@ const permutations = ['full'];
  */
 function closureBuilder(entrypoint) {
   const isRelease = argv.buildtype === 'RELEASE' || argv.release;
+  const isBeta = argv.buildtype === 'BETA' || argv.beta;
   const config = isRelease ? {
     /** -- Release Config -- **/
     'soy': {
@@ -31,8 +31,6 @@ function closureBuilder(entrypoint) {
       'shouldProvideRequireJsFunctions': true
     },
     'closure': {
-      'debug': false,
-
       'warning_level': 'QUIET',
       'dependency_mode': 'STRICT',
       'entry_point': 'goog:bloombox.setup',
@@ -42,14 +40,51 @@ function closureBuilder(entrypoint) {
       'compilation_level': 'ADVANCED',
       'env': 'BROWSER',
       'language_in': 'ECMASCRIPT6',
-      'language_out': 'ECMASCRIPT5_STRICT',
+      'language_out': 'ECMASCRIPT6',
       'export_local_property_definitions': true,
       'generate_exports': true,
       'isolation_mode': 'IIFE',
       'assume_function_wrapper': true,
       'process_closure_primitives': true,
       'rewrite_polyfills': true,
-      'D': 'bloombox.DEBUG=false',
+      'D': ['bloombox.DEBUG=false', 'bloombox.rpc.FALLBACK'],
+      'define': 'bloombox.VERSION=\'' + version + '\'',
+      'create_source_map': 'target/js-' + entrypoint + '.map',
+      'output_module_dependencies': 'target/deps-' + entrypoint + '.json',
+      'source_map_include_content': true,
+      'hide_warnings_for': [
+        'goog/json/json_perf',
+        'goog/storage/mechanism',
+        'closure/goog',
+        'goog'
+      ]
+    }
+  } : (isBeta ? {
+    /** -- Beta Config -- **/
+    'soy': {
+      'useClosureStyle': true,
+      'shouldGenerateJsdoc': true,
+      'shouldProvideRequireSoyNamespaces': true,
+      'shouldProvideRequireJsFunctions': true
+    },
+    'closure': {
+      'warning_level': 'QUIET',
+      'dependency_mode': 'STRICT',
+      'entry_point': 'goog:bloombox.setup',
+      'output_manifest': 'target/manifest-' + entrypoint + '.MF',
+      'charset': 'UTF-8',
+      'use_types_for_optimization': true,
+      'compilation_level': 'ADVANCED',
+      'env': 'BROWSER',
+      'language_in': 'ECMASCRIPT6',
+      'language_out': 'ECMASCRIPT6',
+      'export_local_property_definitions': true,
+      'generate_exports': true,
+      'isolation_mode': 'IIFE',
+      'assume_function_wrapper': true,
+      'process_closure_primitives': true,
+      'rewrite_polyfills': true,
+      'D': ['bloombox.DEBUG=false', 'bloombox.rpc.FALLBACK=false'],
       'define': 'bloombox.VERSION=\'' + version + '\'',
       'create_source_map': 'target/js-' + entrypoint + '.map',
       'output_module_dependencies': 'target/deps-' + entrypoint + '.json',
@@ -70,8 +105,6 @@ function closureBuilder(entrypoint) {
       'shouldProvideRequireJsFunctions': true
     },
     'closure': {
-      'debug': true,
-
       'warning_level': 'DEFAULT',
       'dependency_mode': 'STRICT',
       'entry_point': 'goog:bloombox.setup',
@@ -82,14 +115,14 @@ function closureBuilder(entrypoint) {
       'compilation_level': 'ADVANCED',
       'env': 'BROWSER',
       'language_in': 'ECMASCRIPT6',
-      'language_out': 'ECMASCRIPT5_STRICT',
+      'language_out': 'ECMASCRIPT6',
       'export_local_property_definitions': true,
       'generate_exports': true,
       'isolation_mode': 'IIFE',
       'assume_function_wrapper': true,
       'process_closure_primitives': true,
       'rewrite_polyfills': true,
-      'D': 'bloombox.DEBUG',
+      'D': ['bloombox.DEBUG', 'bloombox.rpc.FALLBACK'],
       'define': 'bloombox.VERSION=\'' + version + '\'',
       'create_source_map': 'target/js-' + entrypoint + '.map',
       'output_module_dependencies': 'target/deps-' + entrypoint + '.json',
@@ -101,7 +134,7 @@ function closureBuilder(entrypoint) {
         'goog'
       ]
     }
-  };
+  });
 
   const externsPath = 'third_party/closure/compiler/contrib/externs/';
   const browserExternsPath = 'third_party/closure/compiler/externs/browser/';
