--- src/menu/v1beta0/section.js
@@ -1,36 +1,0 @@
-
-/*
- * Copyright 2018, Bloombox, LLC.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Bloombox Menu: Sections
- *
- * @fileoverview Enumerates menu sections.
- */
-
-goog.require('proto.opencannabis.products.menu.section.Section');
-
-goog.provide('bloombox.menu.Section');
-
-
-// -- Menu Sections -- //
-/**
- * Menu sections.
- *
- * @export
- * @typedef {proto.opencannabis.products.menu.section.Section}
- */
-bloombox.menu.Section;
