+++ src/shop/v1beta0/ordering/delivery.js
@@ -29,139 +29,144 @@ goog.provide('bloombox.shop.order.DeliveryLocation.fromResponse');
 
 goog.require('bloombox.identity.StreetAddress');
 
+goog.require('bloombox.rpc.FALLBACK');
+
 goog.require('proto.opencannabis.commerce.DeliveryDestination');
 goog.require('proto.opencannabis.geo.Address');
 
 
-// -- Delivery Location -- //
-
-/**
- * Specifies the location for a delivery.
- *
- * @param {bloombox.identity.StreetAddress} streetAddress Delivery address.
- * @param {?string=} instructions Delivery instructions, if any.
- * @constructor
- * @export
- */
-bloombox.shop.order.DeliveryLocation = function DeliveryLocation(streetAddress,
-                                                                 instructions) {
-  /**
-   * Street address for delivery.
-   * @type {bloombox.identity.StreetAddress}
-   */
-  this.streetAddress = streetAddress;
+if (bloombox.rpc.FALLBACK) {
+  // -- Delivery Location -- //
 
   /**
-   * Special instructions for delivery, if any.
-   * @type {?string}
+   * Specifies the location for a delivery.
+   *
+   * @param {bloombox.identity.StreetAddress} streetAddress Delivery address.
+   * @param {?string=} instructions Delivery instructions, if any.
+   * @constructor
+   * @export
    */
-  this.deliveryInstructions = instructions || null;
-};
+  bloombox.shop.order.DeliveryLocation = function DeliveryLocation(streetAddress,
+                                                                   instructions) {
+    /**
+     * Street address for delivery.
+     * @type {bloombox.identity.StreetAddress}
+     */
+    this.streetAddress = streetAddress;
+
+    /**
+     * Special instructions for delivery, if any.
+     * @type {?string}
+     */
+    this.deliveryInstructions = instructions || null;
+  };
 
 
-/**
- * Decode a delivery location from a raw proto object response.
- *
- * @param {?Object} protob Raw response object.
- * @return {?bloombox.shop.order.DeliveryLocation} Decoded delivery location
- *         object, or `null` if none could be decoded.
- */
-bloombox.shop.order.DeliveryLocation.fromResponse = function(protob) {
-  if ((typeof protob === 'object') &&
+  /**
+   * Decode a delivery location from a raw proto object response.
+   *
+   * @param {?Object} protob Raw response object.
+   * @return {?bloombox.shop.order.DeliveryLocation} Decoded delivery location
+   *         object, or `null` if none could be decoded.
+   */
+  bloombox.shop.order.DeliveryLocation.fromResponse = function(protob) {
+    if ((typeof protob === 'object') &&
       (typeof protob['destination'] === 'object')) {
-    // decode location
-    let data = /** @type {Object} */ (protob['destination']);
-    let address = /** @type {Object|undefined} */ (data['address']);
-    let instructions = /** @type {string|null|undefined} */ (
-      data['instructions']);
+      // decode location
+      let data = /** @type {Object} */ (protob['destination']);
+      let address = /** @type {Object|undefined} */ (data['address']);
+      let instructions = /** @type {string|null|undefined} */ (
+        data['instructions']);
 
-    if (typeof address === 'object') {
-      // decode address
-      let firstLine = /** @type {string|undefined|null} */ (
-        address['firstLine']);
-      let secondLine = /** @type {string|undefined|null} */ (
-        address['secondLine']) || null;
-      let city = /** @type {string|undefined|null} */ (
-        address['city']);
-      let state = /** @type {string|undefined|null} */ (
-        address['state']);
-      let zipcode = /** @type {string|undefined|null} */ (
-        address['zipcode']);
+      if (typeof address === 'object') {
+        // decode address
+        let firstLine = /** @type {string|undefined|null} */ (
+          address['firstLine']);
+        let secondLine = /** @type {string|undefined|null} */ (
+          address['secondLine']) || null;
+        let city = /** @type {string|undefined|null} */ (
+          address['city']);
+        let state = /** @type {string|undefined|null} */ (
+          address['state']);
+        let zipcode = /** @type {string|undefined|null} */ (
+          address['zipcode']);
 
-      if (typeof firstLine === 'string' &&
+        if (typeof firstLine === 'string' &&
           typeof city === 'string' &&
           typeof state === 'string' &&
           typeof zipcode === 'string') {
-        // we are good to go
-        return new bloombox.shop.order.DeliveryLocation(
-          new bloombox.identity.StreetAddress(
-            firstLine,
-            secondLine,
-            city,
-            state,
-            zipcode),
-          instructions);
-      } else {
-        // invalid address
-        bloombox.logging.error(
-          'Failed to decode delivery address data.',
-          {'data': data, 'address': address});
+          // we are good to go
+          return new bloombox.shop.order.DeliveryLocation(
+            new bloombox.identity.StreetAddress(
+              firstLine,
+              secondLine,
+              city,
+              state,
+              zipcode),
+            instructions);
+        } else {
+          // invalid address
+          bloombox.logging.error(
+            'Failed to decode delivery address data.',
+            {'data': data, 'address': address});
+        }
       }
     }
-  }
-  return null;
-};
+    return null;
+  };
 
 
-/**
- * Export a proto that specifies this delivery location.
- *
- * @return {proto.opencannabis.commerce.DeliveryDestination} Exported proto.
- * @public
- */
-bloombox.shop.order.DeliveryLocation.prototype.export = function() {
-  let protobuf = new proto.opencannabis.commerce.DeliveryDestination();
-  let address = new proto.opencannabis.geo.Address();
+  /**
+   * Export a proto that specifies this delivery location.
+   *
+   * @return {proto.opencannabis.commerce.DeliveryDestination} Exported proto.
+   * @public
+   */
+  bloombox.shop.order.DeliveryLocation.prototype.export = function() {
+    let protobuf = new proto.opencannabis.commerce.DeliveryDestination();
+    let address = new proto.opencannabis.geo.Address();
 
-  address.setFirstLine(this.streetAddress.firstLine);
-  if (this.streetAddress.secondLine !== null)
-    address.setSecondLine(this.streetAddress.secondLine);
-  address.setCity(this.streetAddress.city);
-  address.setState(this.streetAddress.state);
-  address.setZipcode(this.streetAddress.zip);
-  protobuf.setAddress(address);
+    address.setFirstLine(this.streetAddress.firstLine);
+    if (this.streetAddress.secondLine !== null)
+      address.setSecondLine(this.streetAddress.secondLine);
+    address.setCity(this.streetAddress.city);
+    address.setState(this.streetAddress.state);
+    address.setZipcode(this.streetAddress.zip);
+    protobuf.setAddress(address);
 
-  if (this.deliveryInstructions !== null)
-    protobuf.setInstructions(this.deliveryInstructions);
-  return protobuf;
-};
-
-
-/**
- * Retrieve the street address associated with this delivery location.
- *
- * @return {bloombox.identity.StreetAddress} Delivery street address.
- * @export
- */
-bloombox.shop.order.DeliveryLocation.prototype.getAddress = function() {
-  return this.streetAddress;
-};
+    if (this.deliveryInstructions !== null)
+      protobuf.setInstructions(this.deliveryInstructions);
+    return protobuf;
+  };
 
 
-/**
- * Set special delivery instructions.
- * @param {string} ins Special instructions.
- * @export
- */
-bloombox.shop.order.DeliveryLocation.prototype.setInstructions = function(ins) {
-  this.deliveryInstructions = ins;
-};
+  /**
+   * Retrieve the street address associated with this delivery location.
+   *
+   * @return {bloombox.identity.StreetAddress} Delivery street address.
+   * @export
+   */
+  bloombox.shop.order.DeliveryLocation.prototype.getAddress = function() {
+    return this.streetAddress;
+  };
+
+
+  /**
+   * Set special delivery instructions.
+   * @param {string} ins Special instructions.
+   * @export
+   */
+  bloombox.shop.order.DeliveryLocation.prototype.setInstructions = function(ins) {
+    this.deliveryInstructions = ins;
+  };
 
 // noinspection JSUnusedGlobalSymbols
-/**
- * Clear any special delivery instructions.
- * @export
- */
-bloombox.shop.order.DeliveryLocation.prototype.clearInstructions = function() {
-  this.deliveryInstructions = null;
-};
+  /**
+   * Clear any special delivery instructions.
+   * @export
+   */
+  bloombox.shop.order.DeliveryLocation.prototype.clearInstructions = function() {
+    this.deliveryInstructions = null;
+  };
+
+}
